refactor: reorganize DTO packages for improved clarity and modularization
- Moved response DTOs into dedicated sub-packages for better organization (e.g., `auth`, `medicine`, `labTest`).
- Updated imports across the project to match the new package structure.
- Added placeholder `MedicineServiceImplTest` as part of restructuring.
