+++ src/main/java/com/nexaworks/rafiq/client/Gemini.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
-import com.nexaworks.rafiq.dto.client.RequestBodyDTO;
+import com.nexaworks.rafiq.dto.client.extractDataFromPdf.RequestBodyDTO;
 
 @Component
 @FeignClient(name = "gemini-client", url = "https://generativelanguage.googleapis.com/v1beta/models")
