+++ src/main/java/com/nexaworks/rafiq/config/PostgreSQLFunctionContributor.java
@@ -11,11 +11,9 @@ public class PostgreSQLFunctionContributor implements FunctionContributor {
         var functionRegistry = functionContributions.getFunctionRegistry();
         var typeConfiguration = functionContributions.getTypeConfiguration();
 
-        // Register PostgreSQL full-text search function
         functionRegistry.registerPattern("ts_match", "?1 @@ plainto_tsquery('english', ?2)",
                 typeConfiguration.getBasicTypeRegistry().resolve(StandardBasicTypes.BOOLEAN));
 
-        // Register pg_trgm similarity function
         functionRegistry.registerPattern("similarity", "similarity(?1, ?2)",
                 typeConfiguration.getBasicTypeRegistry().resolve(StandardBasicTypes.DOUBLE));
     }
