+++ src/main/java/com/nexaworks/rafiq/controller/AuthController.java
@@ -12,8 +12,8 @@ import com.nexaworks.rafiq.dto.request.user.ChangePasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.ForgetPasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.ResetPasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.VerifyOtpRequest;
-import com.nexaworks.rafiq.dto.response.LoginResponse;
-import com.nexaworks.rafiq.dto.response.VerifyOtpResponse;
+import com.nexaworks.rafiq.dto.response.auth.LoginResponse;
+import com.nexaworks.rafiq.dto.response.auth.VerifyOtpResponse;
 import com.nexaworks.rafiq.service.AuthService;
 
 import jakarta.servlet.http.HttpServletRequest;
