+++ src/main/java/com/nexaworks/rafiq/controller/SpecializationController.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import com.nexaworks.rafiq.dto.response.SpecializationResponse;
+import com.nexaworks.rafiq.dto.response.specialization.SpecializationResponse;
 import com.nexaworks.rafiq.service.SpecializationService;
 
 import lombok.RequiredArgsConstructor;
