--- src/main/java/com/nexaworks/rafiq/dto/client/TextPart.java
@@ -1,4 +1,0 @@
-package com.nexaworks.rafiq.dto.client;
-
-public record TextPart(String text) {
-}
