+++ src/main/java/com/nexaworks/rafiq/dto/client/extractDataFromPdf/Part.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.client;
+package com.nexaworks.rafiq.dto.client.extractDataFromPdf;
 
 public record Part(InlineDataPart inlineData, String text) {
 }
