+++ src/main/java/com/nexaworks/rafiq/dto/client/extractDataFromPdf/RequestBodyDTO.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.client;
+package com.nexaworks.rafiq.dto.client.extractDataFromPdf;
 
 import java.util.List;
