+++ src/main/java/com/nexaworks/rafiq/dto/client/extractDataFromPdf/TextPart.java
@@ -1,0 +1,4 @@
+package com.nexaworks.rafiq.dto.client.extractDataFromPdf;
+
+public record TextPart(String text) {
+}
