+++ src/main/java/com/nexaworks/rafiq/dto/response/auth/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.auth;
 
 import java.util.Optional;
