+++ src/main/java/com/nexaworks/rafiq/dto/response/auth/VerifyOtpResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.auth;
 
 public record VerifyOtpResponse(String accessToken) {
 }
