+++ src/main/java/com/nexaworks/rafiq/dto/response/labTest/TestResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.labTest;
 
 import java.util.UUID;
