+++ src/main/java/com/nexaworks/rafiq/dto/response/medicine/AddResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.medicine;
 
 public record AddResponse<T>(boolean success, String message, T data) {
 }
