+++ src/main/java/com/nexaworks/rafiq/dto/response/medicine/DrugSearchResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.medicine;
 
 import java.util.UUID;
