+++ src/main/java/com/nexaworks/rafiq/dto/response/socialLinks/SocialLinksResponse.java
@@ -1,4 +1,4 @@
-package com.nexaworks.rafiq.dto.response;
+package com.nexaworks.rafiq.dto.response.socialLinks;
 
 public record SocialLinksResponse(String facebook, String twitter, String instagram,
         String linkedin, String youtube, String whatsapp, String website) {
