+++ src/main/java/com/nexaworks/rafiq/mapper/DrugMapper.java
@@ -3,7 +3,7 @@ package com.nexaworks.rafiq.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-import com.nexaworks.rafiq.dto.response.DrugSearchResponse;
+import com.nexaworks.rafiq.dto.response.medicine.DrugSearchResponse;
 import com.nexaworks.rafiq.entities.Drug;
 
 @Mapper(componentModel = "spring")
