+++ src/main/java/com/nexaworks/rafiq/mapper/ReminderMapper.java
@@ -2,7 +2,7 @@ package com.nexaworks.rafiq.mapper;
 
 import org.mapstruct.Mapper;
 
-import com.nexaworks.rafiq.dto.response.ReminderResponse;
+import com.nexaworks.rafiq.dto.response.reminder.ReminderResponse;
 import com.nexaworks.rafiq.entities.Reminder;
 
 @Mapper(componentModel = "spring")
