+++ src/main/java/com/nexaworks/rafiq/mapper/TestMapper.java
@@ -3,7 +3,7 @@ package com.nexaworks.rafiq.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
-import com.nexaworks.rafiq.dto.response.TestResultsResponse;
+import com.nexaworks.rafiq.dto.response.labTest.TestResultsResponse;
 import com.nexaworks.rafiq.entities.LabTest;
 
 @Mapper(componentModel = "spring")
