+++ src/main/java/com/nexaworks/rafiq/service/ServiceImpl/AuthServiceImpl.java
@@ -21,8 +21,8 @@ import com.nexaworks.rafiq.dto.request.user.ChangePasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.ForgetPasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.ResetPasswordRequest;
 import com.nexaworks.rafiq.dto.request.user.VerifyOtpRequest;
-import com.nexaworks.rafiq.dto.response.LoginResponse;
-import com.nexaworks.rafiq.dto.response.VerifyOtpResponse;
+import com.nexaworks.rafiq.dto.response.auth.LoginResponse;
+import com.nexaworks.rafiq.dto.response.auth.VerifyOtpResponse;
 import com.nexaworks.rafiq.entities.Token;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.GoogleAuthException;
