+++ src/main/java/com/nexaworks/rafiq/service/SpecializationService.java
@@ -3,7 +3,7 @@ package com.nexaworks.rafiq.service;
 import java.util.List;
 import java.util.UUID;
 
-import com.nexaworks.rafiq.dto.response.SpecializationResponse;
+import com.nexaworks.rafiq.dto.response.specialization.SpecializationResponse;
 import com.nexaworks.rafiq.entities.Specialization;
 
 public interface SpecializationService {
