+++ src/main/java/com/nexaworks/rafiq/utils/AuthSessionManager.java
@@ -3,7 +3,7 @@ package com.nexaworks.rafiq.utils;
 import org.springframework.http.ResponseCookie;
 import org.springframework.stereotype.Component;
 
-import com.nexaworks.rafiq.dto.response.LoginResponse;
+import com.nexaworks.rafiq.dto.response.auth.LoginResponse;
 import com.nexaworks.rafiq.entities.Role;
 import com.nexaworks.rafiq.entities.User;
 import com.nexaworks.rafiq.exception.custom.TokenInvalidException;
