+++ src/test/java/com/nexaworks/rafiq/unit/service/MedicineServiceImplTest.java
@@ -1,0 +1,4 @@
+package com.nexaworks.rafiq.unit.service;
+
+public class MedicineServiceImplTest {
+}
